+++ custom_components/metlink_explorer/sensor.py
@@ -305,7 +305,7 @@ class MetlinkRouteSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.6.0",
+            "sw_version": "0.6.1",
         }
 
     @property
@@ -425,7 +425,7 @@ class MetlinkDirectionSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.6.0",
+            "sw_version": "0.6.1",
         }
 
     @property